requested oauth server for refreshing client user token

+++ client/database/migrations/2021_06_21_224554_create_oauth_tokens_table.php
@@ -16,7 +16,7 @@ class CreateOauthTokensTable extends Migration
         Schema::create('oauth_tokens', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained();
-            $table->integer('expires_in');
+            $table->bigInteger('expires_in');
             $table->text('access_token');
             $table->text('refresh_token');
             $table->timestamps();
